copy legacy examples to deprecated tab
delete examples from legacy dir

add todo

bump

rename TableComposable demos and examples

fix filename

revert md references for now

fix name

fix a11y config

test update paths

revert

fix duplicate id

revert path update

remove whitespace
