+++ packages/react-table/src/components/TableComposable/examples/TableActionsOverflow.tsx
@@ -24,7 +24,7 @@ interface Repository {
   isMenuOpen: boolean;
 }
 
-export const ComposableTableActions: React.FunctionComponent = () => {
+export const TableActions: React.FunctionComponent = () => {
   // In real usage, this data would come from some external source like an API via props.
   const repositories: Repository[] = [
     { name: 'one', branches: 'two', prs: 'a', workspaces: 'four', lastCommit: 'five', isMenuOpen: false },
