+++ packages/react-table/src/components/TableComposable/examples/TableMultipleStickyColumns.tsx
@@ -14,7 +14,7 @@ interface Fact {
   detail7: string;
 }
 
-export const ComposableTableMultipleStickyColumns: React.FunctionComponent = () => {
+export const TableMultipleStickyColumns: React.FunctionComponent = () => {
   // In real usage, this data would come from some external source like an API via props.
   const facts: Fact[] = Array.from({ length: 9 }, (_, index) => ({
     name: `Fact ${index + 1}`,
