+++ packages/react-table/src/components/TableComposable/examples/TableSortable.tsx
@@ -9,7 +9,7 @@ interface Repository {
   lastCommit: string;
 }
 
-export const ComposableTableSortable: React.FunctionComponent = () => {
+export const TableSortable: React.FunctionComponent = () => {
   // In real usage, this data would come from some external source like an API via props.
   const repositories: Repository[] = [
     { name: 'one', branches: 'two', prs: 'a', workspaces: 'four', lastCommit: 'five' },
