+++ packages/react-table/src/components/TableComposable/examples/TableSortableCustom.tsx
@@ -13,7 +13,7 @@ interface Repository {
   lastCommit: string;
 }
 
-export const ComposableTableSortableCustom: React.FunctionComponent = () => {
+export const TableSortableCustom: React.FunctionComponent = () => {
   // In real usage, this data would come from some external source like an API via props.
   const repositories: Repository[] = [
     { name: 'one', branches: 'two', prs: 'a', workspaces: 'four', lastCommit: 'five' },
