+++ packages/react-table/src/components/TableComposable/examples/TableStickyColumnsAndHeader.tsx
@@ -24,7 +24,7 @@ interface Fact {
   detail7: string;
 }
 
-export const ComposableTableStickyColumnsAndHeader: React.FunctionComponent = () => {
+export const TableStickyColumnsAndHeader: React.FunctionComponent = () => {
   // In real usage, this data would come from some external source like an API via props.
   const facts: Fact[] = Array.from({ length: 9 }, (_, index) => ({
     name: `Fact ${index + 1}`,
